chore: folder restructure and css styling for listings page
