+++ client/src/components/UserProfile/UserProfile.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button } from 'react-bootstrap';
-import { userProfileStyle, userpicStyle } from '../styles/customStyle.js';
+import { userProfileStyle, userpicStyle } from '../../utils/customStyle.js';
 var fakeUserpic = 'http://www.lovemarks.com/wp-content/uploads/profile-avatars/default-avatar-rainbow-unicorn.png';
 
 export default class UserProfile extends React.Component {
