+++ client/src/utils/Utils.jsx
@@ -1,6 +1,7 @@
 import $ from 'jquery';
 
 exports.requireAuth = (nextState, replace) => {
+  //Is user logged in?
   if (false /*user is not logged in*/) {
     console.log('nope, you cant do that without logging in');
     replace('/login');
